+++ examples/main/shared/viewer/Viewer.jsx
@@ -5,8 +5,14 @@ import { isSSR } from 'wix-rich-content-common';
 import * as PropTypes from 'prop-types';
 import * as Plugins from './ViewerPlugins';
 import theme from '../theme/theme'; // must import after custom styles
-import getImagesData from 'wix-rich-content-fullscreen/src/lib/getImagesData';
+import getImagesData from 'wix-rich-content-fullscreen/dist/lib/getImagesData';
 import Fullscreen from 'wix-rich-content-fullscreen';
+import 'wix-rich-content-fullscreen/dist/styles.min.css';
+
+import {
+  TextSelectionListener,
+  ViewerInlineToolBar,
+} from 'wix-rich-content-text-selection-toolbar';
 import { GALLERY_TYPE } from 'wix-rich-content-plugin-gallery';
 const anchorTarget = '_top';
 const relValue = 'noreferrer';
@@ -38,20 +44,17 @@ export default class Viewer extends PureComponent {
     onExpand: (entityIndex, innerIndex = 0) => {
       //galleries have an innerIndex (i.e. second image will have innerIndex=1)
       this.setState({
-        expendModeIsOpen: true,
+        expandModeIsOpen: true,
         expandModeIndex: this.expandModeData.imageMap[entityIndex] + innerIndex,
       });
     },
   };
 
   render() {
-    const { isMobile, locale, initialState, seoMode } = this.props;
-    const { expendModeIsOpen, expandModeIndex, disabled } = this.state;
+    const { isMobile, initialState, seoMode } = this.props;
+    const { expandModeIsOpen, expandModeIndex, disabled } = this.state;
 
     const viewerProps = {
-      locale,
-      relValue,
-      anchorTarget,
       isMobile,
       theme,
       initialState,
@@ -60,25 +63,30 @@ export default class Viewer extends PureComponent {
     };
 
     return (
-      <div id="rich-content-viewer" className="viewer">
-        <RichContentViewer
-          helpers={this.helpers}
-          typeMappers={Plugins.typeMappers}
-          inlineStyleMappers={Plugins.getInlineStyleMappers(initialState)}
-          decorators={Plugins.decorators}
-          config={this.pluginsConfig}
-          // siteDomain="https://www.wix.com"
-          {...viewerProps}
-        />
-        {!isSSR() && (
-          <Fullscreen
-            isOpen={expendModeIsOpen}
-            images={this.expandModeData.images}
-            onClose={() => this.setState({ expendModeIsOpen: false })}
-            index={expandModeIndex}
+      <>
+        <div id="rich-content-viewer" className="viewer">
+          <RichContentViewer
+            helpers={this.helpers}
+            typeMappers={Plugins.typeMappers}
+            inlineStyleMappers={Plugins.getInlineStyleMappers(initialState)}
+            decorators={Plugins.decorators}
+            config={this.pluginsConfig}
+            // siteDomain="https://www.wix.com"
+            {...viewerProps}
           />
+          {!isSSR() && (
+            <Fullscreen
+              images={this.expandModeData.images}
+              onClose={() => this.setState({ expandModeIsOpen: false })}
+              isOpen={expandModeIsOpen}
+              index={expandModeIndex}
+            />
+          )}
+        </div>
+        {!isSSR() && (
+          <TextSelectionListener targetId={'rich-content-viewer'} ToolBar={ViewerInlineToolBar} />
         )}
-      </div>
+      </>
     );
   }
 }