+++ examples/storybook/.storybook/webpack.config.js
@@ -1,24 +1,5 @@
 const wixStorybookConfig = require('yoshi/config/webpack.config.storybook');
 
 module.exports = ({ config }) => {
-  const baseConfig = wixStorybookConfig(config);
-  return {
-    ...baseConfig,
-    module: {
-      ...baseConfig.module,
-      rules: [
-        {
-          test: /\.m?js$/,
-          exclude: /(node_modules|bower_components)/,
-          use: {
-            loader: 'babel-loader',
-            options: {
-              plugins: ['babel-plugin-preval'],
-            },
-          },
-        },
-        ...baseConfig.module.rules,
-      ],
-    },
-  };
+  return wixStorybookConfig(config);
 };