+++ packages/editor-common/web/src/Utils/draftUtils.js
@@ -1,5 +1,5 @@
 import { EditorState, Modifier, RichUtils, SelectionState, AtomicBlockUtils } from 'draft-js';
-import { cloneDeep, flatMap, findIndex, findLastIndex } from 'lodash';
+import { cloneDeep, flatMap, findIndex, findLastIndex, countBy } from 'lodash';
 
 function createSelection({ blockKey, anchorOffset, focusOffset }) {
   return SelectionState.createEmpty(blockKey).merge({
@@ -59,18 +59,35 @@ function isSelectionBelongsToExsistingLink(editorState, selection) {
   });
 }
 
+function preventLinkInlineStyleForNewLine(editorState, { anchorKey, focusOffset }) {
+  const selectionForSpace = createSelection({
+    blockKey: anchorKey,
+    anchorOffset: focusOffset,
+    focusOffset,
+  });
+  //insert dummy space after link for preventing underline inline style to the new line
+  return Modifier.insertText(editorState.getCurrentContent(), selectionForSpace, ' ');
+}
+
 function insertLink(editorState, selection, data) {
   const oldSelection = editorState.getSelection();
-  const newContentState = Modifier.applyInlineStyle(
-    addEntity(editorState, selection, {
-      type: 'LINK',
-      data: createLinkEntityData(data),
-    }).getCurrentContent(),
-    selection,
-    'UNDERLINE'
-  ).set('selectionAfter', oldSelection);
+  const editorWithLink = addEntity(editorState, selection, {
+    type: 'LINK',
+    data: createLinkEntityData(data),
+  });
+  const isNewLine = selection.anchorKey !== oldSelection.anchorKey; //check weather press enter or space after link
+  const contentState = isNewLine
+    ? preventLinkInlineStyleForNewLine(editorWithLink, selection)
+    : editorWithLink.getCurrentContent();
 
-  return EditorState.push(editorState, newContentState, 'change-inline-style');
+  return EditorState.push(
+    editorState,
+    Modifier.applyInlineStyle(contentState, selection, 'UNDERLINE').set(
+      'selectionAfter',
+      oldSelection
+    ),
+    'change-inline-style'
+  );
 }
 
 function createLinkEntityData({ url, targetBlank, nofollow, anchorTarget, relValue }) {
@@ -306,7 +323,7 @@ function getSelectedLinksInBlock(block, editorState) {
     }));
 }
 
-function getLinkRangesInBlock(block, contentState) {
+export function getLinkRangesInBlock(block, contentState) {
   const ranges = [];
   block.findEntityRanges(
     value => {
@@ -352,6 +369,70 @@ function getSelection(editorState) {
   return selection;
 }
 
+export const getEntities = (editorState, entityType = null) => {
+  const currentContent = editorState.getCurrentContent();
+  const entities = [];
+
+  currentContent.getBlockMap().forEach(block => {
+    block.findEntityRanges(character => {
+      const char = character.getEntity();
+      const entity = !!char && currentContent.getEntity(char);
+      if (!entityType || entity.getType() === entityType) {
+        entities.push(entity);
+      }
+    });
+  });
+  return entities;
+};
+
+const countByType = obj => countBy(obj, x => x.type);
+
+const getBlockTypePlugins = blocks =>
+  blocks.filter(block => block.type !== 'unstyled' && block.type !== 'atomic');
+
+export function getPostContentSummary(editorState) {
+  if (Object.entries(editorState).length === 0) return;
+  const blocks = editorState.getCurrentContent().getBlocksAsArray();
+  const entries = getEntities(editorState);
+  const blockPlugins = getBlockTypePlugins(blocks);
+  return {
+    postContent: {
+      ...countByType(blockPlugins),
+      ...countByType(entries),
+    },
+  };
+}
+
+//ATM, looks for deleted plugins.
+//onChanges - for phase 2?
+//Added Plugins - checked elsewhere via toolbar clicks
+export const calculateDiff = async (prevState, newState, onPluginDelete) => {
+  const countByType = obj => countBy(obj, x => x.type);
+  const prevEntities = countByType(getEntities(prevState));
+  const currEntities = countByType(getEntities(newState));
+  const prevBlocks = prevState.getCurrentContent().getBlocksAsArray();
+  const currBlocks = newState.getCurrentContent().getBlocksAsArray();
+  const prevBlockPlugins = countByType(getBlockTypePlugins(prevBlocks));
+  const currBlockPlugins = countByType(getBlockTypePlugins(currBlocks));
+
+  const prevPluginsTotal = Object.assign(prevEntities, prevBlockPlugins);
+  const currPluginsTotal = Object.assign(currEntities, currBlockPlugins);
+
+  Object.keys(prevPluginsTotal).forEach(type => {
+    if (!currPluginsTotal[type] || prevPluginsTotal[type] > currPluginsTotal[type]) {
+      onPluginDelete(type);
+    }
+  });
+
+  // onPluginChange -> for Phase 2
+  //else {
+  // const before = beforePlugins[key];
+  // const after = afterPlugins[key];
+  // if (JSON.stringify(before) !== JSON.stringify(after))
+  //   onPluginChange(type, { from: before, to: after });
+  //}
+};
+
 // a selection of the new content from the last change
 function createLastChangeSelection(editorState) {
   const content = editorState.getCurrentContent();
@@ -391,9 +472,9 @@ export function getBlockInfo(editorState, blockKey) {
   const contentState = editorState.getCurrentContent();
   const block = contentState.getBlockForKey(blockKey);
   const entityKey = block.getEntityAt(0);
-  const entity = entityKey && contentState.getEntity(entityKey);
-  const entityData = entity?.data;
-  const type = entity?.type;
+  const entity = (entityKey && contentState.getEntity(entityKey)) || {};
+  const entityData = entity.data;
+  const type = entity.type;
 
   return { type: type || 'text', entityData };
 }