+++ packages/editor-common/web/src/Utils/resizeDecorationPropHandlers.js
@@ -4,7 +4,7 @@ export const pluginDecorationProps = (props, componentData) => {
   const resizeableProps = PLUGIN_DECORATION_PROPS[PLUGIN_DECORATIONS.RESIZEABLE](props);
   const decorationProps = { ...resizeableProps };
   if (componentData) {
-    const { size } = componentData.config;
+    const size = componentData.config?.size;
     const isInlineSize = size === 'inline';
     const { width, ...rest } = resizeableProps.style; // eslint-disable-line no-unused-vars
     const style = isInlineSize ? resizeableProps.style : { ...rest };